function_discord_bot/requirements.txt
pynacl
flask
google-cloud-storage